requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==23.2.0
blinker==1.7.0
braintree==4.28.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.2
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
PyJWT==2.8.0
pymongo==4.6.3
python-dotenv==1.0.1
requests==2.31.0
twilio==9.0.4
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
waitress==3.0.0
